+++ desktop_app/src/ui/user_management.py
@@ -7,7 +7,6 @@ from desktop_app.src.ui.components import StyledButton
 from public_api.api import UsersAPI, APIClient
 from public_api.shared_schemas import UserSanitizedWithRole, UserFilter, AllPermissions, UserWithPermissions, \
     UserCreate, UserUpdate, AllRoles
-from public_api.shared_schemas.user import RoleName
 
 
 class UserManagementWidget(QWidget):
@@ -19,6 +18,7 @@ class UserManagementWidget(QWidget):
         self.users_api = UsersAPI(api_client)
         self.roles = self.users_api.get_all_roles().roles
         self.role_name_to_id = {role.role_name: role.id for role in self.roles}
+        self.permission_manager = self.users_api.get_current_user_permissions()
         self.init_ui()
 
     def init_ui(self):
@@ -62,9 +62,10 @@ class UserManagementWidget(QWidget):
         self.stacked_widget.addWidget(main_widget)
 
         # Floating Action Button for adding new users
-        self.fab = StyledButton("+")
-        self.fab.clicked.connect(self.add_user)
-        layout.addWidget(self.fab)
+        if self.permission_manager.has_write_permission("user_management"):
+            self.fab = StyledButton("+")
+            self.fab.clicked.connect(self.add_user)
+            layout.addWidget(self.fab)
 
         self.refresh_users()
 
@@ -83,7 +84,6 @@ class UserManagementWidget(QWidget):
             self.table.setItem(row, 2, QTableWidgetItem(user.role.role_name))
             self.table.setItem(row, 3, QTableWidgetItem("Active" if user.is_active else "Inactive"))
 
-            # Convert timestamp to QDateTime
             if user.last_login is not None:
                 last_login_dt = QDateTime.fromSecsSinceEpoch(user.last_login)
                 last_login_str = last_login_dt.toString("yyyy-MM-dd HH:mm:ss")
@@ -94,18 +94,21 @@ class UserManagementWidget(QWidget):
             actions_widget = QWidget()
             actions_layout = QHBoxLayout(actions_widget)
             actions_layout.setContentsMargins(0, 0, 0, 0)
-            actions_layout.setSpacing(2)  # Reduce spacing between buttons
+            actions_layout.setSpacing(2)
 
-            edit_button = StyledButton("Edit")
-            edit_button.clicked.connect(lambda _, uid=user.id: self.edit_user(uid))
-            permissions_button = StyledButton("Permissions")
-            permissions_button.clicked.connect(lambda _, uid=user.id: self.manage_permissions(uid))
-            delete_button = StyledButton("Delete")
-            delete_button.clicked.connect(lambda _, uid=user.id: self.delete_user(uid))
+            if self.permission_manager.has_write_permission("user_management"):
+                edit_button = StyledButton("Edit")
+                edit_button.clicked.connect(lambda _, uid=user.id: self.edit_user(uid))
+                actions_layout.addWidget(edit_button)
 
-            actions_layout.addWidget(edit_button)
-            actions_layout.addWidget(permissions_button)
-            actions_layout.addWidget(delete_button)
+                permissions_button = StyledButton("Permissions")
+                permissions_button.clicked.connect(lambda _, uid=user.id: self.manage_permissions(uid))
+                actions_layout.addWidget(permissions_button)
+
+            if self.permission_manager.has_delete_permission("user_management"):
+                delete_button = StyledButton("Delete")
+                delete_button.clicked.connect(lambda _, uid=user.id: self.delete_user(uid))
+                actions_layout.addWidget(delete_button)
 
             self.table.setCellWidget(row, 5, actions_widget)
 
@@ -116,7 +119,6 @@ class UserManagementWidget(QWidget):
         for row in range(self.table.rowCount()):
             row_match = True
 
-            # Text search
             if search_text:
                 row_match = False
                 for col in range(self.table.columnCount()):
@@ -125,7 +127,6 @@ class UserManagementWidget(QWidget):
                         row_match = True
                         break
 
-            # Role filtering
             if selected_role_id is not None and row_match:
                 role_item = self.table.item(row, 2)
                 if role_item:
